feat: add python CI
