+++ .gitignore
@@ -1,2 +1,4 @@
 html
 __pycache__
+poetry-installer-error-*.log
+docs/
